requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2
feedparser
pythonping
pynewtonmath
PyLyrics
hurry.filesize
tldextract
PyYAML
pyowm
wikipedia
bleach
markdown2
speedtest-cli>=2.0.2
bs4
lxml
googletrans>=2.4.0
